backend/auth_service/requirements.txt

fastapi
uvicorn
python-jose
passlib[bcrypt]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
